Add the header for CFConvNeighbors
